% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{explore_data}
\alias{explore_data}
\title{Explore the dataset}
\usage{
explore_data(setup, type = c("visual", "tabular"), factors = NULL, by = NULL)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{type}{Character scalar. Either \code{visual} or \code{tabular}.}

\item{factors}{Character vector. Names of selected potential predictors to limit the scope of results.}

\item{by}{Character scalar. Name of potential predictor that will constitute the second dimension.}
}
\value{
Either list of dataframes or list of ggplot2 charts.
}
\description{
Helps to explore data in either visual or tabular form. The view can also be one or two-dimensional.
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

explore_data(setup, type = "tabular")
explore_data(setup, type = "visual")
explore_data(setup, type = "tabular", by = "pol_yr")
explore_data(setup, type = "visual", by = "pol_yr")

}
\seealso{
\code{\link{explore_target}}, \code{\link{explore_corr}}
}
