% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_target.R
\name{explore_target}
\alias{explore_target}
\title{Explore the target variable}
\usage{
explore_target(
  setup,
  type = c("visual", "tabular"),
  weighted = TRUE,
  exclude_zero = FALSE,
  lower_quantile = 0,
  upper_quantile = 1,
  n_cuts = 20
)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{type}{Character scalar. Either \code{visual} or \code{tabular}.}

\item{weighted}{Logical scalar. Whether the distribution should be weighted.}

\item{exclude_zero}{Logical scalar. Whether to exclude records with zero value of target variable.}

\item{lower_quantile}{Numeric scalar. Lower quantile to include in the distribution analysis.}

\item{upper_quantile}{Numeric scalar. Upper quantile to include in the distribution analysis.}

\item{n_cuts}{Integer scalar. Number of cuts in the printed distribution. More cuts means more detailed view.}
}
\value{
Either a ggplot2 chart or vector of values printed to console.
}
\description{
Helps to explore the distribution of target variable in either visual or tabular form.
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

explore_target(setup)
explore_target(setup, type = 'tabular')
explore_target(setup, type = 'visual')
explore_target(setup, type = 'tabular', exclude_zero = TRUE)
explore_target(setup, type = 'visual', lower_quantile = 0.05, upper_quantile = 0.95)

}
\seealso{
\code{\link{explore_data}}, \code{\link{explore_corr}}
}
