% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_add.R
\name{factor_add}
\alias{factor_add}
\title{Add predictor to the current model formula}
\usage{
factor_add(setup, var_symbol)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{var_symbol}{Unquoted symbol. Predictor to be added. Must be present in the modeling dataset.}
}
\value{
Setup object with updated attributes.
}
\description{
Adds the predictor to the list of predictors. These will be used in model formula when \code{model_fit} is called.
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

print(setup)

# not fitted yet
modeling <- setup \%>\%
  factor_add(pol_yr) \%>\%
  factor_add(agecat)

print(modeling)


}
