% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_remove.R
\name{factor_remove}
\alias{factor_remove}
\title{Remove predictor from the current model formula}
\usage{
factor_remove(setup, var_symbol)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{var_symbol}{Unquoted symbol. Predictor to be removed. Must be present in the modeling dataset.}
}
\value{
Setup object with updated attributes.
}
\description{
Removes the predictor from the list of predictors. These will be used in model formula when \code{model_fit} is called.
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = sev_train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

modeling <- setup \%>\%
  factor_add(pol_yr) \%>\%
  factor_add(agecat) \%>\%
  model_fit() \%>\%
  factor_remove(agecat)

}
