% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_betas.R
\name{model_betas}
\alias{model_betas}
\title{Inspect insuRglm model coefficients}
\usage{
model_betas(setup, triangles = FALSE)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{triangles}{Boolean scalar. Show beta coefficient differences instead. These are expressed using standard error percentage.}
}
\value{
Either dataframe containing model coefficients or matrix-like dataframe containing beta coefficient differences
}
\description{
Inspect insuRglm model coefficients
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

modeling <- setup \%>\%
  factor_add(pol_yr) \%>\%
  factor_add(agecat) \%>\%
  model_fit()

modeling \%>\%
  model_betas()

modeling \%>\%
  model_betas(triangles = TRUE)

}
