% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_crossval.R
\name{model_crossval}
\alias{model_crossval}
\title{Train all insuRglm models on CV data}
\usage{
model_crossval(setup, cv_folds = 10, stratified = FALSE)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{cv_folds}{Integer scalar. Number of rancom CV folds to be used.}

\item{stratified}{Boolean scalar. Whether to stratify losses and non-losses.
This will help in creating more representative crossvalidation folds with datasets that contain very few non-zero losses.}
}
\value{
Setup object with updated attributes.
}
\description{
Train the current (last) and any saved insuRglm model using CV data. Predictions are stored for later use.
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = sev_train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

modeling <- setup \%>\%
  factor_add(pol_yr) \%>\%
  factor_add(agecat) \%>\%
  model_fit()

modeling_cv <- modeling \%>\%
  model_crossval()

modeling_cv \%>\%
  model_lift(data = 'crossval')

}
\seealso{
\code{\link{model_save}}, \code{\link{model_lift}}, \code{\link{model_performance}}
}
