% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_export.R
\name{model_export}
\alias{model_export}
\title{Export the current insuRglm model}
\usage{
model_export(setup, xlsx_file, overwrite = FALSE)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{xlsx_file}{Character scalar. Path and name of the file to store the exported model.}

\item{overwrite}{Boolean scalar. Whether to overwrite an existing file.}
}
\value{
NULL, this function has only side effects.
}
\description{
Exports the current model into xlsx file. The excel spreadsheet will contain the charts, as well as
relativities and weights for each predictor included in the current (last) model.
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

modeling <- setup \%>\%
  factor_add(pol_yr) \%>\%
  factor_add(agecat) \%>\%
  model_fit()

modeling \%>\%
  model_export('export_test.xlsx')

}
