% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{Fit the insuRglm model}
\usage{
model_fit(setup)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}
}
\value{
Setup object with updated attributes.
}
\description{
Fits the model with the current model formula. Computes and saves back many new attributes and objects.
This is a required step before using \code{model_visualize}, \code{model_compare}, \code{model_save},
\code{model_betas}, \code{model_crossval}, \code{model_lift}, \code{model_export} and others
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = sev_train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

modeling <- setup \%>\%
  factor_add(pol_yr) \%>\%
  factor_add(agecat) \%>\%
  model_fit()

modeling \%>\%
  model_visualize(factors = 'fitted')

modeling \%>\%
  model_visualize(factors = 'unfitted')

}
