% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lift.R
\name{model_lift}
\alias{model_lift}
\title{Visualize the lift of insuRglm models}
\usage{
model_lift(
  setup,
  data = c("train", "crossval"),
  model = c("current", "all"),
  buckets = 10,
  weighted = TRUE
)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{data}{Character scalar. Either \code{train} or \code{crossval}. The latter requires running \code{model_crossval} first.}

\item{model}{Character scalar. Either \code{current} or \code{all}. The latter will display the current and all saved models.}

\item{buckets}{Integer scalar. Number of groups to divide data into.}

\item{weighted}{Boolean scalar. Whether the average of target variable in each group should be weighted.}
}
\value{
List of one or more ggplo2 charts.
}
\description{
Visualize the lift chart of one or all saved insuRglm models. The records are first ordered by predictions in ascending order and
then divided roughly into several buckets (groups). Average of target variable is computed and displayed for each group
separately.
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

modeling <- setup \%>\%
  factor_add(pol_yr) \%>\%
  factor_add(agecat) \%>\%
  model_fit()

modeling \%>\%
  model_lift(data = 'train', buckets = 10)

modeling_cv <- modeling \%>\%
  model_crossval()

modeling_cv \%>\%
  model_lift(data = 'crossval', buckets = 5)

}
\seealso{
\code{\link{model_save}}, \code{\link{model_crossval}}
}
