% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_revert.R
\name{model_revert}
\alias{model_revert}
\title{Revert back to a specific insuRglm model}
\usage{
model_revert(setup, to)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{to}{Character scalar. Name of the saved model that we wish to revert to.}
}
\value{
Setup object with updated attributes.
}
\description{
Reverts back to a specific saved insuRglm model within the setup (modeling) object.
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

modeling <- setup \%>\%
  factor_add(pol_yr) \%>\%
  factor_add(agecat) \%>\%
  model_fit() \%>\%
  model_save('model1') \%>\%
  factor_modify(agecat = variate(agecat, type = 'non_prop', mapping = c(1, 2, 3, 4, 5,6))) \%>\%
  model_fit() \%>\%
  model_revert(to = 'model1') # from now on the two lines above have no effect (but they stay documented)

}
\seealso{
\code{\link{model_save}}
}
