% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_save.R
\name{model_save}
\alias{model_save}
\title{Save fitted insuRglm model for later use}
\usage{
model_save(setup, name = NULL)
}
\arguments{
\item{setup}{Setup object. Created at the start of the workflow. Usually piped in from previous step.}

\item{name}{Character scalar. Model will be saved under this name.}
}
\value{
Setup object with updated attributes.
}
\description{
Saves already fitted insuRglm model for later use, which might be comparison using \code{model_compare} or
returning back to a previous model using \code{model_revert}.
}
\examples{
require(dplyr) # for the pipe operator
data('sev_train')

setup <- setup(
  data_train = sev_train,
  target = 'sev',
  weight = 'numclaims',
  family = 'gamma',
  keep_cols = c('pol_nbr', 'exposure', 'premium')
)

modeling <- setup \%>\%
  factor_add(pol_yr) \%>\%
  factor_add(agecat) \%>\%
  model_fit() \%>\%
  model_save('model1') \%>\%
  factor_modify(agecat = variate(agecat, type = 'non_prop', mapping = c(1, 2, 3, 4, 5, 6))) \%>\%
  model_fit()

modeling \%>\%
  model_compare(with = 'model1')

}
\seealso{
\code{\link{model_compare}}, \code{\link{model_revert}}
}
